requirements.txt
djangorestframework>=3.6
sqlparse>=0.2.4
python-dateutil>=2.6.1
